refactor to more readable directories
